@@ -1,7 +1,7 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import {
-  FormGroup,
-  FormBuilder,
+  UntypedFormGroup,
+  UntypedFormBuilder,
   Validators,
   FormGroupDirective,
 } from '@angular/forms';
@@ -29,7 +29,7 @@ export class QueryDetailsComponent implements OnInit {
   promoExistErr = false;
   promocodeId = '';
   PromocodeCreatedOn = '';
-  queryForm: FormGroup;
+  queryForm: UntypedFormGroup;
   promocodeObj: any = {};
   tenantList: any;
   siteList: any;
@@ -59,7 +59,7 @@ export class QueryDetailsComponent implements OnInit {
   constructor(
     public filterService: GridFilterService,
     private cdref: ChangeDetectorRef,
-    private readonly formBuilder: FormBuilder,
+    private readonly formBuilder: UntypedFormBuilder,
     private popUpService: PopUpService,
     private httpDataService: HttpDataService,
     public dialog: MatDialog,
