@@ -9,7 +9,7 @@ import { HttpDataService } from '@app/shared/services/http-data.service';
 import { RouterExtService } from '@app/shared/services/routerExt.service';
 import { AppConstants } from '@app/constants';
 import { PopUpService } from '@app/shared/utility/popup.service';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { Observable } from 'rxjs';
 import Helper from 'src/app/shared/utility/Helper';
 import { GridFilterService } from '@app/shared/utility/grid-filter.service';
@@ -31,15 +31,15 @@ export class RegisterChargePointComponent implements OnInit {
   siteName: string;
   deletedRecords = false;
   dialogRef: MatDialogRef<any>;
-  chargePointIdControl = new FormControl();
+  chargePointIdControl = new UntypedFormControl();
   filteredByChargePointId: Observable<string[]>;
   chargePointId: any[];
   chargePointIdValues: any[];
-  chargeTypeControl = new FormControl();
+  chargeTypeControl = new UntypedFormControl();
   filteredByChargeType: Observable<string[]>;
   chargeTypeValues: any[];
   chargeType: any[];
-  statusControl = new FormControl();
+  statusControl = new UntypedFormControl();
   filteredByStatus: Observable<string[]>;
   statusValues: any[];
   status: any[];
